test: split Playwright workflow helpers
Keep Playwright fixture seams in harness.ts and move user-facing app flows to workflows.ts.

Update specs to import harness assertions separately from app workflow helpers.
