+++ tests/integration/fs/fs-empty.spec.ts
@@ -1,4 +1,5 @@
-import { connectFS, expect, test } from '../helpers/harness.ts';
+import { expect, test } from '../helpers/harness.ts';
+import { connectFS } from '../helpers/workflows.ts';
 
 test('connecting to an empty sync folder leaves the library empty', async ({ page }) => {
   await connectFS(page);
